move build_stream_transfer_data to model_executor/utils
